+++ routes/turnoRoute.js
@@ -4,7 +4,6 @@ const api = express.Router();
 const turnoController = require('../controllers/turnoController');
 
 api.post('/', turnoController.createTurno);
-api.post('/mailer/', turnoController.sendEmail);
 api.get('/', turnoController.getTurno);
 api.get('/:id', turnoController.getSpecific);
 api.put('/set/:id', turnoController.setEntrada);
